C6 (#113): split features/exams/dal.ts → student/admin/shared (pur déplacement, baril compat)
1518 lignes → dal.student.ts (12 fonctions getCurrentSession), dal.admin.ts
(6 fonctions requireRole admin), dal.shared.ts (helpers images/comptage +
getOpenExam*QuestionIds + types pont). dal.ts devient un baril de ré-export :
aucun import existant ne casse. Fixe au passage l'inférence du ternaire dans
callAction d'exam-form.tsx (union createExam/updateExam non assignable,
masquée par le cache incrémental tsc au gate précédent).
